backend/requirements.txt
numpy
opencv-python
pysstv
psutil
pyadi-iio
pillow
fastapi
uvicorn
smbus2
requests